mm-audio: aenc: make frame_count variable 64 bit
- After 71 minutes AAC encoding stops working
- The root cause is while calculating the timestamp
  of the frame a variable frame_count is 32 bit and
  it ovewrflows which leads incorrect timestamp
  calculation which leads to stopping of AAC encoding
- To solve this make frame_count 64 bitvariable.

CRs-Fixed: 707449

+++ mm-audio/aenc-aac/qdsp6/inc/omx_aac_aenc.h
@@ -457,7 +457,7 @@ private:
     unsigned int                   m_flags;      //encapsulate the waiting states.
     OMX_U64                        nTimestamp;
     OMX_U64                        ts;
-    uint32_t                       m_frame_count;
+    uint64_t                       m_frame_count;
     unsigned int                   frameduration;
     unsigned int                   pcm_input; //tunnel or non-tunnel
     unsigned int                   m_inp_act_buf_count;    // Num of Input Buffers
